tagging/requirements.txt
kafka-python==2.0.2
numpy<2
torch>=2.3.0
transformers>=4.44.2
psycopg2-binary==2.9.9
pydantic
sqlalchemy
pydantic-settings
pymysql